% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransLasso-functions.R
\name{agg.fun}
\alias{agg.fun}
\title{Title}
\usage{
agg.fun(B, X.test, y.test, total.step = 10, selection = F)
}
\arguments{
\item{B}{matrix}

\item{X.test}{matrix}

\item{y.test}{vector}

\item{total.step}{integer}

\item{selection}{logical}
}
\value{
list
}
\description{
Title
}
