% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{approx_marginal_likelihood}
\alias{approx_marginal_likelihood}
\title{Compute the marginal likelihood}
\usage{
approx_marginal_likelihood(
  y0,
  yA,
  yAb,
  X0,
  XA,
  XAb,
  d_A,
  d_Ab,
  d_delta,
  cut1 = 0.1
)
}
\description{
Compute the marginal likelihood
}
