% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{construct_candidate_set}
\alias{construct_candidate_set}
\title{Construct candidate sets from sparsity index (Li et al., 2020)}
\usage{
construct_candidate_set(X, y, n.vec)
}
\description{
Construct candidate sets from sparsity index (Li et al., 2020)
}
