% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransLasso-functions.R
\name{las.kA}
\alias{las.kA}
\title{Title}
\usage{
las.kA(X, y, A0, n.vec, lam.const = NULL, l1 = T)
}
\arguments{
\item{X}{design matrix}

\item{y}{response vector}

\item{A0}{initial set}

\item{n.vec}{vector of sample sizes}

\item{lam.const}{lambda constant}

\item{l1}{logical indicating l1 or l0 method}
}
\value{
list
}
\description{
Title
}
