% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{precompute_XtXXty}
\alias{precompute_XtXXty}
\title{Precompute XtX and Xty for all source studies}
\usage{
precompute_XtXXty(X, y, K, n.vec)
}
\description{
Precompute XtX and Xty for all source studies
}
