% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_multistudy_regression.R
\name{simulate_multistudy_regression}
\alias{simulate_multistudy_regression}
\title{Generates data for high-dimensional linear regression}
\usage{
simulate_multistudy_regression(
  p,
  s,
  M,
  size.A0,
  n.vec,
  sig.beta,
  sig.delta1,
  sig.delta2,
  beta_bias,
  contam_pct = 0.015,
  ni_contam_factor = NULL,
  test_set = FALSE,
  pct_test = NULL,
  type = "gaussian"
)
}
\arguments{
\item{p}{number of predictors}

\item{s}{number of signals}

\item{M}{number of studies}

\item{size.A0}{number of informative studies}

\item{n.vec}{vector of sample sizes for each study}

\item{sig.beta}{signal size}

\item{beta_bias}{signal bias}

\item{contam_pct}{contamination percentage}

\item{ni_contam_factor}{non-informative contamination factor}

\item{test_set}{whether to generate a test set}

\item{pct_test}{percentage of data to use for test set}

\item{type}{type of data to generate}
}
\description{
Generates data for high-dimensional linear regression
}
