% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransLasso-functions.R
\name{Trans.lasso}
\alias{Trans.lasso}
\title{Title}
\usage{
Trans.lasso(X, y, n.vec, I.til, l1 = T)
}
\arguments{
\item{X}{design matrix}

\item{y}{response vector}

\item{n.vec}{vector of sample sizes}

\item{I.til}{holdout set index}

\item{l1}{logical indicating l1 or l0 method}
}
\value{
list
}
\description{
Title
}
