% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast_select.R
\name{blast_select}
\alias{blast_select}
\title{Bayesian Transfer Learning Shrinkage Regression with Source Selection}
\usage{
blast_select(
  X,
  y,
  n.vec,
  burn = 1000,
  iter = 3000,
  a = 1/5,
  b = 10,
  s = 0.8,
  tau = 1,
  sigma2 = 1,
  w = 1,
  alpha = 0.05,
  iterEBstep = 0,
  sir = TRUE,
  gamma_init = NULL,
  temp_scale = 1,
  adapt_temp_scale = FALSE
)
}
\arguments{
\item{X}{predictor matrix}

\item{y}{response vector}

\item{n.vec}{vector of sample sizes for each study}

\item{burn}{burn-in iterations}

\item{iter}{number of iterations}

\item{a}{parameter for rejection sampler}

\item{b}{parameter for rejection sampler}

\item{s}{tuning parameter for proposal distribution}

\item{tau}{global shrinkage parameter}

\item{sigma2}{error variance}

\item{w}{inverse gamma prior shape parameter}

\item{alpha}{confidence level for credible intervals}

\item{iterEBstep}{number of iterations for empirical bayes step, if 0, no empirical bayes}

\item{sir}{whether to use SIR for empirical bayes}
}
\description{
Bayesian Transfer Learning Shrinkage Regression with Source Selection
}
