% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horseshoe_empirical_bayes.R
\name{eb_em_max}
\alias{eb_em_max}
\title{Empirical Bayes maximization via EM algorithm}
\usage{
eb_em_max(
  kappa,
  N,
  xi_out,
  sigma2_out,
  iterEBstep,
  eb_counter,
  i,
  sir = F,
  maxiter = 20,
  tol = 1e-06
)
}
\arguments{
\item{kappa}{initial value of global shrinkage prior parameter}

\item{N}{total number of observations}

\item{xi_out}{global shrinkage parameter draws from posterior}

\item{sigma2_out}{error variance draws from posterior}

\item{iterEBstep}{number of iterations for empirical bayes step}

\item{eb_counter}{counter for empirical bayes step}

\item{i}{current iteration}

\item{sir}{whether to use SIR for empirical bayes}

\item{maxiter}{maximum number of iterations for EB step}

\item{tol}{tolerance for EB step}
}
\description{
Empirical Bayes maximization via EM algorithm
}
