% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horseshoe_gibbs_step.R
\name{exact_horseshoe_gibbs_step}
\alias{exact_horseshoe_gibbs_step}
\title{One step of the Gibbs sampler for the horseshoe model}
\usage{
exact_horseshoe_gibbs_step(
  X,
  y,
  eta,
  xi,
  Q,
  w,
  s,
  a,
  b,
  p,
  sigma2,
  sigma2_0 = NULL,
  kappa,
  n0 = NULL,
  N_k = NULL,
  EB,
  p_star = NULL,
  kappa_0_trans = FALSE,
  kappa_A_bar_trans = FALSE,
  c0 = 1/2,
  c_bar = 1,
  xiA_trunc = FALSE,
  xiA_trunc_val = NULL
)
}
\arguments{
\item{X}{Predictor matrix}

\item{y}{Response vector}

\item{eta}{current value of local shrinkage parameter}

\item{xi}{current value of global shrinkage parameter}

\item{Q}{XtX}

\item{w}{inverse gamma prior shape parameter}

\item{s}{tuning parameter for the proposal distribution}

\item{a}{parameter for the rejection sampler}

\item{b}{parameter for the rejection sampler}

\item{p}{number of predictors}

\item{sigma2}{error variance}

\item{kappa}{parameter for global shrinkage prior}

\item{EB}{logical, whether to use empirical Bayes or not}

\item{p_star}{number of non-zero coefficients if want to specify}

\item{kappa_0_trans}{logical for whether to transform overall sparsity of bias parameters for TL}

\item{kappa_A_bar_trans}{logical for whether to transform overall sparsity of noninformative studies for TL}

\item{c0}{factor for number of signals for bias}

\item{c_bar}{factor for number of signals for noninformative studies}
}
\description{
One step of the Gibbs sampler for the horseshoe model
}
